+++ src/main/java/draylar/tiered/reforge/widget/ReforgeTab.java
@@ -3,11 +3,14 @@ package draylar.tiered.reforge.widget;
 import org.jetbrains.annotations.Nullable;
 
 import draylar.tiered.network.TieredClientPacket;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.libz.api.InventoryTab;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 
+@Environment(EnvType.CLIENT)
 public class ReforgeTab extends InventoryTab {
 
     public ReforgeTab(Text title, @Nullable Identifier texture, int preferedPos, Class<?>... screenClasses) {